Changement de structure afin d'avoir plusieurs fenêtre en fonction des fonctionnalités 22/02/2021
