+++ src/main/java/sample/Controller.java
@@ -59,11 +59,11 @@ import java.util.stream.IntStream;
 
 
 public class Controller implements Initializable {
-    private ObservableList<Bibliotheque.Livre> livres = FXCollections.observableArrayList();
-    private int Livreindex;
+    protected ObservableList<Bibliotheque.Livre> livres = FXCollections.observableArrayList();
+    protected int Livreindex;
     private File selectedFile;
     @FXML private javafx.scene.control.MenuItem CloseAppButton;
-    @FXML private TableView<Bibliotheque.Livre> tableBook;
+    @FXML protected TableView<Bibliotheque.Livre> tableBook;
     @FXML private TableColumn<Bibliotheque.Livre, String> TitreColumn;
     @FXML private TableColumn<Bibliotheque.Livre, String> AuteurColumn;
     @FXML private TableColumn<Bibliotheque.Livre, String> ResumeColumn;
@@ -77,11 +77,10 @@ public class Controller implements Initializable {
     @FXML private TextField ColonneInput;
     @FXML private TextField RangeeInput;
     @FXML private TextArea ResumeInput;
-    @FXML private TextField URLInput;
     @FXML private RadioButton pret;
     @FXML private RadioButton available;
     @FXML private ImageView bookURL;
-
+    @FXML private AjoutController ajoutController;
 
     /**
      * Cette méthode permet d'initialiser l'interface ainsi que notre tableau et notre event sur celui-ci.
@@ -122,7 +121,6 @@ public class Controller implements Initializable {
                     else{
                         available.setSelected(true);
                     }
-                    URLInput.setText(rowData.getURL());
                     showBookImage(rowData.getURL());
                 }
             });
@@ -153,12 +151,6 @@ public class Controller implements Initializable {
         }
         bookURL.setImage(image);
     }
-    public void unselectPret(){
-        pret.setSelected(false);
-    }
-    public void unselectDispo(){
-        available.setSelected(false);
-    }
     /**
      * Cette méthode permet d'afficher le menu de confirmation de l'arrêt de l'application.
      * @param event sert à executer la méthode se trouvant dans le onAction de notre fichier sample.fxml.
@@ -171,6 +163,19 @@ public class Controller implements Initializable {
             Stage stage = new Stage();
             stage.setScene(new Scene(root1));
             stage.setTitle("Quitter");
+            stage.show();
+
+        } catch (Exception e) {
+            System.out.println("raté");
+        }
+    }
+    public void showAjoutLivre(ActionEvent event) {
+        try {
+            FXMLLoader fxmlLoader = new FXMLLoader(getClass().getResource("/fxml/ajout.fxml"));
+            Parent root1 = (Parent)fxmlLoader.load();
+            Stage stage = new Stage();
+            stage.setScene(new Scene(root1));
+            stage.setTitle("Ajout d'un Livre");
             stage.show();
 
         } catch (Exception e) {
@@ -197,73 +202,6 @@ public class Controller implements Initializable {
     }
 
     /**
-     * Cette méthode permet de valider le formulaire, elle respecte le constructeur livre
-     * et si des informations manquent, elle remplit automatiquement les champs.
-     * A terme cette méthode permettra d'envoyer les nouvelles données dans le tableau
-     * @param Event C'est l'évenement de cliquer sur le bouton
-     */
-    @FXML
-    private void validerLivre (ActionEvent Event){
-        String prenom, nom= "";
-        String titre = TitreInput.getText();
-        String res=ResumeInput.getText();
-        String aut=AuteurInput.getText();
-        String url=URLInput.getText();
-        try{
-            int c =Integer.parseInt(ColonneInput.getText());
-            int paru=Integer.parseInt(ParutionInput.getText());
-            int r =Integer.parseInt(RangeeInput.getText());
-            if (c<=5 && c>=1 && r<=7 && r>=1){
-                if (TitreInput.getText().isEmpty()){
-                    titre= "Titre incconu";
-                }
-                if (AuteurInput.getText().indexOf(" ")==-1){
-                    if (AuteurInput.getText().isEmpty()){
-                        aut="auteur inconnu";
-                    }
-                    else{
-                        aut=" "+AuteurInput.getText();
-                    }
-                }
-                if (ResumeInput.getText().isEmpty()){
-                    res="résumé vide";
-                }
-                String[] auteur= aut.split(" ");
-                prenom=auteur[0];
-                nom=auteur[1];
-                Bibliotheque.Livre l1 = new Bibliotheque.Livre();
-                Bibliotheque.Livre.Auteur auteur1 = new Bibliotheque.Livre.Auteur();
-                auteur1.setNom(nom);
-                auteur1.setPrenom(prenom);
-                l1.setAuteur(auteur1);
-                l1.setTitre(titre);
-                l1.setColonne((short)c);
-                l1.setParution(paru);
-                l1.setPresentation(res);
-                l1.setRangee((short) r);
-                if(pret.isSelected()){
-                    l1.setEtat("En Prêt");
-                }
-                else if(available.isSelected()){
-                    l1.setEtat("Disponible");
-                }
-                l1.setURL(url);
-
-                tableBook.setItems(getLivre(l1));
-                disableInput();
-                resetInput();
-            }
-            else{
-                erreur();
-            }
-        }
-        catch(NumberFormatException e){
-            erreur();
-        }
-
-    }
-
-    /**
      * Permet l'ouverture de l'explorateur de fichier afin de choisir un fichier XML
      * @param actionEvent ce paramètre permet l'action de cliquer sur le bouton et d'excuter la méthode.
      */
@@ -282,18 +220,6 @@ public class Controller implements Initializable {
     /**
      * Affiche un message d'erreur lorsque les valeurs limites de colonne ou rangée
      */
-    @FXML
-    public void erreur() {
-        try {
-            FXMLLoader fxmlLoader = new FXMLLoader(getClass().getResource("/fxml/Erreur.fxml"));
-            Parent root1 = (Parent) fxmlLoader.load();
-            Stage stage = new Stage();
-            stage.setTitle("Erreur");
-            stage.setScene(new Scene(root1));
-            stage.show();
-        }
-        catch (Exception e){System.out.println("raté");}
-    }
 
     /**
      * Permet d'empêcher l'écriture dans les inputs tant que l'utilisateur n'a pas appuyer sur le bouton Ajouter
@@ -307,7 +233,6 @@ public class Controller implements Initializable {
         ParutionInput.setDisable(true);
         pret.setDisable(true);
         available.setDisable(true);
-        URLInput.setDisable(true);
     }
 
     /**
@@ -322,7 +247,6 @@ public class Controller implements Initializable {
         ParutionInput.setDisable(false);
         pret.setDisable(false);
         available.setDisable(false);
-        URLInput.setDisable(false);
         resetInput();
     }
 
@@ -338,69 +262,10 @@ public class Controller implements Initializable {
         ResumeInput.setText("");
         pret.setSelected(false);
         available.setSelected(false);
-        URLInput.setText("");
     }
     public void suppLivre(){
         livres.remove(Livreindex);
         resetInput();
-    }
-    public void modifLivre(){
-        Bibliotheque.Livre l = livres.get(Livreindex);
-        String prenom, nom= "";
-        String titre = TitreInput.getText();
-        String res=ResumeInput.getText();
-        String aut=AuteurInput.getText();
-        String url=URLInput.getText();
-        try{
-            int c =Integer.parseInt(ColonneInput.getText());
-            int paru=Integer.parseInt(ParutionInput.getText());
-            int r =Integer.parseInt(RangeeInput.getText());
-            if (c<=5 && c>=1 && r<=7 && r>=1){
-                if (TitreInput.getText().isEmpty()){
-                    titre= "Titre incconu";
-                }
-                if (AuteurInput.getText().indexOf(" ")==-1){
-                    if (AuteurInput.getText().isEmpty()){
-                        aut="auteur inconnu";
-                    }
-                    else{
-                        aut=" "+AuteurInput.getText();
-                    }
-                }
-                if (ResumeInput.getText().isEmpty()){
-                    res="résumé vide";
-                }
-                String[] auteur= aut.split(" ");
-                prenom=auteur[0];
-                nom=auteur[1];
-                Bibliotheque.Livre.Auteur auteur1 = new Bibliotheque.Livre.Auteur();
-                auteur1.setNom(nom);
-                auteur1.setPrenom(prenom);
-                l.setAuteur(auteur1);
-                l.setTitre(titre);
-                l.setColonne((short)c);
-                l.setParution(paru);
-                l.setPresentation(res);
-                l.setRangee((short) r);
-                if(pret.isSelected()){
-                    l.setEtat("En Prêt");
-                }
-                else if(available.isSelected()){
-                    l.setEtat("Disponible");
-                }
-                l.setURL(url);
-                livres.set(Livreindex,l);
-                tableBook.setItems(getLivre2(l));
-                disableInput();
-                resetInput();
-            }
-            else{
-                erreur();
-            }
-        }
-        catch(NumberFormatException e){
-            erreur();
-        }
     }
 
     /**
