+++ src/main/resources/fxml/ajout.fxml
@@ -9,11 +9,11 @@
 <?import javafx.scene.layout.BorderPane?>
 <?import javafx.scene.layout.Pane?>
 
-<Pane maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="451.0" prefWidth="281.0" xmlns="http://javafx.com/javafx/15.0.1" xmlns:fx="http://javafx.com/fxml/1">
+<Pane maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="451.0" prefWidth="281.0" xmlns="http://javafx.com/javafx/11.0.1" xmlns:fx="http://javafx.com/fxml/1" fx:controller="sample.AjoutController">
    <children>
       <BorderPane prefHeight="448.0" prefWidth="301.0">
          <left>
-            <AnchorPane prefHeight="448.0" prefWidth="283.0" BorderPane.alignment="CENTER">
+            <AnchorPane fx:id="ajoutLivre" prefHeight="445.0" prefWidth="305.0" BorderPane.alignment="CENTER">
                <children>
                   <TextField fx:id="TitreInput" layoutX="108.0" layoutY="14.0" />
                   <TextField fx:id="AuteurInput" layoutX="108.0" layoutY="50.0" />
